+++ test/base/src/androidTest/kotlin/io/realm/test/shared/QueryTests.kt
@@ -16,11 +16,11 @@
 
 package io.realm.test.shared
 
-import io.realm.QuerySort
 import io.realm.Realm
 import io.realm.RealmConfiguration
 import io.realm.entities.Sample
-import io.realm.find
+import io.realm.query.Sort
+import io.realm.query.find
 import io.realm.test.platform.PlatformUtils
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
@@ -53,7 +53,7 @@ class QueryTests {
     }
 
     @Test
-    fun find() {
+    fun find_realm() {
         realm.query(Sample::class)
             .find { results -> assertEquals(0, results.size) }
 
@@ -81,7 +81,34 @@ class QueryTests {
     }
 
     @Test
-    fun find_malformedQueryThrows() {
+    fun find_mutableRealm() {
+        val stringValue = "some string"
+
+        realm.writeBlocking {
+            // Queries inside a write transaction are live and can be reused within the closure
+            val query = query(Sample::class)
+
+            query.find { results -> assertEquals(0, results.size) }
+
+            assertEquals(0, query.find().size)
+            copyToRealm(Sample().apply { stringField = stringValue })
+            assertEquals(1, query.find().size)
+
+            // No filter
+            query.find { results -> assertEquals(1, results.size) }
+
+            // Filter by string
+            query.query("stringField = $0", stringValue)
+                .find { results -> assertEquals(1, results.size) }
+
+            // Filter by string that doesn't match
+            query.query("stringField = $0", "invalid string")
+                .find { results -> assertEquals(0, results.size) }
+        }
+    }
+
+    @Test
+    fun find_realm_malformedQueryThrows() {
         assertFailsWith<IllegalArgumentException> {
             realm.query(Sample::class, "stringField = $0")
         }.let {
@@ -102,6 +129,29 @@ class QueryTests {
     }
 
     @Test
+    fun find_mutableRealm_malformedQueryThrows() {
+        realm.writeBlocking {
+            assertFailsWith<IllegalArgumentException> {
+                query(Sample::class, "stringField = $0")
+            }.let {
+                assertTrue(it.message!!.contains("Have you specified all parameters"))
+            }
+
+            assertFailsWith<IllegalArgumentException> {
+                query(Sample::class, "stringField = 42")
+            }.let {
+                assertTrue(it.message!!.contains("Wrong query field"))
+            }
+
+            assertFailsWith<IllegalArgumentException> {
+                query(Sample::class, "nonExistingField = 13")
+            }.let {
+                assertTrue(it.message!!.contains("Wrong query field"))
+            }
+        }
+    }
+
+    @Test
     @Ignore
     fun asFlow() {
         // TODO
@@ -114,60 +164,186 @@ class QueryTests {
     }
 
     @Test
+    @SuppressWarnings("LongMethod")
     fun composedQuery() {
-        val value1 = 1
-        val value2 = 2
-        val value3 = 3
-        val value4 = 4
+        val joe = "Joe"
+        val sylvia = "Sylvia"
+        val stacy = "Stacy"
+        val ruth = "Ruth"
+        val bob = "Bob"
 
         realm.writeBlocking {
-            copyToRealm(Sample().apply { intField = value1 })
-            copyToRealm(Sample().apply { intField = value2 })
-            copyToRealm(Sample().apply { intField = value3 })
-            copyToRealm(Sample().apply { intField = value4 })
+            copyToRealm(
+                Sample().apply {
+                    intField = 1
+                    stringField = joe
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = 2
+                    stringField = sylvia
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = 3
+                    stringField = stacy
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = 4
+                    stringField = ruth
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = 5
+                    stringField = bob
+                }
+            )
         }
 
+        // 5, 4, 3
         realm.query(Sample::class, "intField > 1")
             .query("intField > 2")
-            .find()
-            .let { results ->
+            .sort(Sample::intField.name, Sort.DESCENDING)
+            .find { results ->
+                assertEquals(3, results.size)
+                assertEquals(5, results[0].intField)
+                assertEquals(4, results[1].intField)
+                assertEquals(3, results[2].intField)
+            }
+
+        // Sylvia, Stacy
+        realm.query(Sample::class, "intField > 1")
+            .query("stringField BEGINSWITH[c] $0", "S")
+            .sort(Sample::intField.name, Sort.ASCENDING)
+            .find { results ->
                 assertEquals(2, results.size)
+                assertEquals(sylvia, results[0].stringField)
+                assertEquals(stacy, results[1].stringField)
+            }
+
+        // Ruth
+        realm.query(Sample::class)
+            .query("stringField BEGINSWITH[c] $0 OR stringField BEGINSWITH[c] $1", "J", "R")
+            .query("intField > 1")
+            .find { results ->
+                assertEquals(1, results.size)
+                assertEquals(ruth, results[0].stringField)
             }
     }
 
     @Test
+    @SuppressWarnings("LongMethod")
     fun composedQuery_withDescriptor() {
         val value1 = 1
         val value2 = 2
         val value3 = 3
         val value4 = 4
+        val value5 = 5
+        val joe = "Joe"
+        val sylvia = "Sylvia"
+        val stacy = "Stacy"
+        val ruth = "Ruth"
+        val bob = "Bob"
 
         realm.writeBlocking {
-            copyToRealm(Sample().apply { intField = value1 })
-            copyToRealm(Sample().apply { intField = value2 })
-            copyToRealm(Sample().apply { intField = value3 })
-            copyToRealm(Sample().apply { intField = value4 })
+            copyToRealm(
+                Sample().apply {
+                    intField = value1
+                    stringField = joe
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value2
+                    stringField = sylvia
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value3
+                    stringField = stacy
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value4
+                    stringField = ruth
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value5
+                    stringField = bob
+                }
+            )
         }
 
-        realm.query(Sample::class, "intField > 1")
+        realm.query(Sample::class)
             .query("intField > 2")
-            .sort(Sample::intField.name, QuerySort.DESCENDING)
+            .sort(Sample::intField.name, Sort.DESCENDING)
             .find()
             .let { results ->
-                assertEquals(2, results.size)
-                assertEquals(value4, results[0].intField)
-                assertEquals(value3, results[1].intField)
+                assertEquals(3, results.size)
+                assertEquals(value5, results[0].intField)
+                assertEquals(value4, results[1].intField)
+                assertEquals(value3, results[2].intField)
             }
 
-        realm.query(Sample::class, "intField > 1")
+        realm.query(Sample::class)
             .query("intField > 2")
-            .sort(Sample::intField.name, QuerySort.DESCENDING)
+            .sort(Sample::intField.name, Sort.DESCENDING)
             .limit(1)
             .find()
             .let { results ->
                 assertEquals(1, results.size)
-                assertEquals(value4, results[0].intField)
+                assertEquals(value5, results[0].intField)
             }
+
+        // Descriptor in query string - Bob, Ruth, Stacy
+        realm.query(Sample::class)
+            .query("intField > 2 SORT(stringField ASCENDING)")
+            .find()
+            .let { results ->
+                assertEquals(3, results.size)
+                assertEquals(bob, results[0].stringField)
+                assertEquals(ruth, results[1].stringField)
+                assertEquals(stacy, results[2].stringField)
+            }
+
+        // Descriptor as a function - Bob, Ruth, Stacy
+        realm.query(Sample::class)
+            .query("intField > 2")
+            .sort(Sample::stringField.name, Sort.ASCENDING)
+            .find()
+            .let { results ->
+                assertEquals(3, results.size)
+                assertEquals(bob, results[0].stringField)
+                assertEquals(ruth, results[1].stringField)
+                assertEquals(stacy, results[2].stringField)
+            }
+
+        // Conflicting descriptors, query string vs function - Bob, Ruth, Stacy
+        realm.query(Sample::class)
+            .query("intField > 2 SORT(stringField ASCENDING)")
+            .sort(Sample::stringField.name, Sort.DESCENDING)
+            .find()
+            .let { results ->
+                assertEquals(3, results.size)
+                assertEquals(bob, results[2].stringField)
+                assertEquals(ruth, results[1].stringField)
+                assertEquals(stacy, results[0].stringField)
+            }
+
+        // Invalid descriptor in query string
+        assertFailsWith<IllegalArgumentException> {
+            realm.query(Sample::class)
+                .query("intField > 2 SORT(stringField DESCENDINGGGGGGGGGG)")
+        }
     }
 
     @Test
@@ -176,22 +352,64 @@ class QueryTests {
         val value2 = 2
         val value3 = 3
         val value4 = 4
+        val sylvia = "Sylvia"
+        val stacy = "Stacy"
+        val ruth = "Ruth"
 
         realm.writeBlocking {
-            copyToRealm(Sample().apply { intField = value1 })
-            copyToRealm(Sample().apply { intField = value2 })
-            copyToRealm(Sample().apply { intField = value3 })
-            copyToRealm(Sample().apply { intField = value4 })
+            copyToRealm(
+                Sample().apply {
+                    intField = value1
+                    stringField = sylvia
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value2
+                    stringField = sylvia // intentionally repeated
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value3
+                    stringField = stacy
+                }
+            )
+            copyToRealm(
+                Sample().apply {
+                    intField = value4
+                    stringField = ruth
+                }
+            )
         }
 
-        realm.query(Sample::class, "intField > 1")
-            .query("intField > 2")
-            .sort(Sample::intField.name, QuerySort.DESCENDING)
-            .query("intField < 4")
-            .find()
+        // Ruth 4, Stacy 3
+        realm.query(Sample::class)
+            .distinct(Sample::stringField.name) // Sylvia, Stacy, Ruth
+            .sort(Sample::stringField.name, Sort.ASCENDING) // Ruth, Stacy, Sylvia
+            .limit(2) // Ruth 4, Stacy 3
+            .find { results ->
+                assertEquals(2, results.size)
+                assertEquals(ruth, results[0].stringField)
+                assertEquals(stacy, results[1].stringField)
+            }
+
+        // Stacy 3, Sylvia 1
+        // One would expect this to be Stacy 3 but notice the last 'query', it moves the descriptors
+        // to the end - this is how it is also implemented in Java
+        realm.query(Sample::class)
+            .distinct(Sample::stringField.name) // normal execution: Sylvia 1, Stacy 3, Ruth 4
+            .sort(
+                Sample::stringField.name,
+                Sort.ASCENDING
+            ) // normal execution: Ruth 4, Stacy 4, Sylvia 1
+            .limit(2) // normal execution: Ruth 4, Stacy 3
+            .query("intField < $value4") // this puts all the previous descriptors at the end of this query!
+            .find() // results for: "intField < 4 DISTINCT('stringField')"
             .let { results ->
-                assertEquals(1, results.size)
-                assertEquals(value3, results[0].intField)
+                assertEquals(2, results.size)
+                assertEquals(stacy, results[0].stringField)
+                assertEquals(sylvia, results[1].stringField)
             }
     }
 
@@ -237,7 +455,7 @@ class QueryTests {
 
         // No filter, sort descending
         realm.query(Sample::class)
-            .sort(Sample::intField.name to QuerySort.DESCENDING)
+            .sort(Sample::intField.name to Sort.DESCENDING)
             .find { results ->
                 assertEquals(3, results.size)
                 assertEquals(ruth.first, results[0].intField)
@@ -270,8 +488,8 @@ class QueryTests {
         // No filter, multiple sortings
         realm.query(Sample::class)
             .sort(
-                Sample::stringField.name to QuerySort.DESCENDING,
-                Sample::intField.name to QuerySort.ASCENDING
+                Sample::stringField.name to Sort.DESCENDING,
+                Sample::intField.name to Sort.ASCENDING
             ).find { results ->
                 assertEquals(3, results.size)
                 assertEquals(ruth.first, results[0].intField)
@@ -309,7 +527,7 @@ class QueryTests {
         }
 
         realm.query(Sample::class)
-            .sort(Sample::intField.name, QuerySort.DESCENDING)
+            .sort(Sample::intField.name, Sort.DESCENDING)
             .limit(2)
             .find { results ->
                 assertEquals(2, results.size)
@@ -317,7 +535,7 @@ class QueryTests {
             }
 
         realm.query(Sample::class)
-            .sort(Sample::intField.name, QuerySort.DESCENDING)
+            .sort(Sample::intField.name, Sort.DESCENDING)
             .limit(2)
             .limit(1)
             .find { results ->